+++ lotus_backend/src/main/java/isamrs/tim17/lotus/repository/NurseRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import isamrs.tim17.lotus.model.Nurse;
 
-public interface NurseRepository extends JpaRepository<Nurse, Integer> {
+public interface NurseRepository extends JpaRepository<Nurse, Long> {
 	
 	Nurse findOneById(String id);
 	Page<Nurse> findAll(Pageable pageable);
