+++ lotus_backend/src/main/java/isamrs/tim17/lotus/service/NurseService.java
@@ -12,7 +12,7 @@ public class NurseService {
 	@Autowired
 	private NurseRepository nurses;
 	
-	public Nurse findOne(int id) {
+	public Nurse findOne(long id) {
 		return nurses.findById(id).orElseGet(null);
 	}
 
@@ -24,7 +24,7 @@ public class NurseService {
 		return nurses.save(doctor);
 	}
 
-	public void remove(int id) {
+	public void remove(long id) {
 		nurses.deleteById(id);
 	}
